+++ account_invoice_reminder_sms_46elks/__manifest__.py
@@ -1,6 +1,6 @@
 {
     "name": "Betalningspåminnelse som SMS via 46elks",
-    "version": "19.0.1.0.0",
+    "version": "19.0.1.0.1",
     "category": "Accounting",
     "summary": "Sättet 'SMS via 46elks' för betalningspåminnelser",
     "author": "Molnkontakt AB",